requirements.txt
discord.py
yt-dlp
asyncio
pynacl
python-dotenv
uuid